Related issue: https://github.com/radareorg/cutter/issues/1765

`time radare2 -A -qq -Q libcrackme.so`

Before:
```
real	1m49,847s
user	1m19,119s
sys	0m30,339s
```
After:
```
real	0m37,498s
user	0m36,761s
sys	0m0,715s
```

But `r_core_anal_cc_init()` still spends some cpu time on switching between 16/32 bits. Data caching can give ~5% analysis speedup (for the given binary).