@ITAYC0HEN 2019-10-18
this PR ready to be merged, right? a massive improvement.
only tests to fix?