@radare 2019-10-18
Lovely. I was aware of this issue. Thanks for addressing it

> On 17 Oct 2019, at 22:53, Paul I. <notifications@github.com> wrote:
> 
> ﻿
> Related issue: radareorg/cutter#1765
> 
> time radare2 -A -qq -Q libcrackme.so
> 
> Before:
> 
> real	1m49,847s
> user	1m19,119s
> sys	0m30,339s
> After:
> 
> real	0m37,498s
> user	0m36,761s
> sys	0m0,715s
> But r_core_anal_cc_init() still spends some cpu time on switching between 16/32 bits. Data caching can give ~5% analysis speedup (for the given binary).
> 
> You can view, comment on, or merge this pull request online at:
> 
>   https://github.com/radareorg/radare2/pull/15306
> 
> Commit Summary
> 
> Avoid reloading of cc sdb
> File Changes
> 
> M libr/core/cbin.c (11)
> Patch Links:
> 
> https://github.com/radareorg/radare2/pull/15306.patch
> https://github.com/radareorg/radare2/pull/15306.diff
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub, or unsubscribe.